backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
cryptography==45.0.6
httpx==0.28.1
openai==1.107.2
langsmith==0.4.13
python-frontmatter==1.1.0
PyYAML==6.0.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
pytest==8.4.1

